Make width and height of a window unreal
Before this commit width and height of an USER object were "real", i.e.
include all pixel on a dimension. For example:

+----+----+----+----+
| 10 | 11 | 12 | 13 |
+----+----+----+----+

Let's assume the draw from above shows horizontal line of pixels where
'10' is coordinate of left point and '13' is coordinate of right point.
Both ten and thirteen are pixels. The previous code gave you '4' as
width of the line. And that's a real width.

But a common technique for Win32 API is to use width and height not
included one of the end edges. Therefore, width of the line should be
'13 - 10 = 3 px'. And this commit calculates like this.

I do not know if this is relevant to .NET, but it is for native Windows
32 API. MSFT Spy++ utility also showed incorrect width and height for
windows from the previous code version.

See also: #13