Set limits on the zoom factors in circle chart
To avoid GDI+ errors this commit defines the limit on minimal available
value of a zoom factor. To avoid floating-point overflow errors it also
enables the upper limit on zoom factors.

This commit changes scaling hotkeys to OEM plus and minus, and plus and
minus keys on the "additional" keypad.

Signed-off-by: Ruslan Garipov <brigadir15@gmail.com>