Remove some hotkeys ambiguity
Before this commit's modifications, the "Ancestors tree" and
"Descendants tree" both used keys `+' and `-' on the keypad to change
width of segmants. But their base class also used the same keys to
change current zoom factor. As a result of such duplicated keys
processing, chart scale and sectors width grew or decreased
simultaneously.

This commit forces user to use Shift+'+' or Shift+'-' to control
chart's segments width, and no additional keys to scale a chart.

Signed-off-by: Ruslan Garipov <brigadir15@gmail.com>